+++ opentelemetry-exporter-geneva/geneva-uploader/src/lib.rs
@@ -18,7 +18,9 @@ pub(crate) use ingestion_service::uploader::{
 };
 
 pub use client::EncodedBatch;
-pub use client::{AgentFedCredential, AgentFedCredentialFuture, AgentFedCredentialSource};
+pub use client::{
+    AccountRouting, AgentFedCredential, AgentFedCredentialFuture, AgentFedCredentialSource,
+};
 pub use client::{
     GenevaClient, GenevaClientConfig, LogsConfig, LogsEventNameMapping, LogsEventNameRoutingKey,
     SpanEventNameMapping, SpanEventNameRoutingKey, TracesConfig, UploadError,